+++ tests/ccl_test_correlation_3dRSD.c
@@ -2,7 +2,7 @@
 #include "ctest.h"
 #include <stdlib.h>
 #include <stdio.h>
-#include <math.h> 
+#include <math.h>
 
 CTEST_DATA(corrs_3dRSD) {
   double Omega_c;
@@ -65,7 +65,7 @@ static void compare_correlation_3dRSD(int i_model,struct corrs_3dRSD_data * data
   FILE *f;
   ccl_configuration config = default_config;
   config.matter_power_spectrum_method= ccl_halofit;
-  config.transfer_function_method = ccl_boltzmann;
+  config.transfer_function_method = ccl_boltzmann_class;
   ccl_parameters params = ccl_parameters_create(data->Omega_c,data->Omega_b,data->Omega_k[i_model-1],
 		data->Neff, data->mnu, data->mnu_type, data->w_0[i_model-1],data->w_a[i_model-1],
 		data->h,data->A_s,data->n_s,-1, -1, -1, -1,NULL,NULL, &status);
@@ -73,8 +73,8 @@ static void compare_correlation_3dRSD(int i_model,struct corrs_3dRSD_data * data
   params.Omega_l=data->Omega_v[i_model-1];
   params.sigma8=data->sigma8;
   ccl_cosmology * cosmo = ccl_cosmology_create(params, config);
-  ASSERT_NOT_NULL(cosmo);      
-  
+  ASSERT_NOT_NULL(cosmo);
+
   sprintf(fname,"./tests/benchmark/model%d_xiRSD.txt",i_model);
   f=fopen(fname,"r");
   if(f==NULL) {
@@ -92,7 +92,7 @@ static void compare_correlation_3dRSD(int i_model,struct corrs_3dRSD_data * data
   double *r_arr1=malloc(N1*sizeof(double));
   double *r_arr2=malloc((nr-N1)*sizeof(double));
   double *ximm_bench_arr=malloc(nr*sizeof(double));
-  
+
   rtn = fgets(str, 1024, f);
   for(i=0;i<nr;i++) {
     double r_h;
@@ -108,7 +108,7 @@ static void compare_correlation_3dRSD(int i_model,struct corrs_3dRSD_data * data
     else
       r_arr2[i-N1]=r_h;
 
-    double ximm_bench;      
+    double ximm_bench;
     stat=fscanf(f,"%lf",&ximm_bench);
     if(stat!=1) {
 	fprintf(stderr,"Error reading file %s, line %d\n",fname,i+2);
@@ -121,30 +121,30 @@ static void compare_correlation_3dRSD(int i_model,struct corrs_3dRSD_data * data
   double *ximm_ccl_out1=malloc(N1*sizeof(double));
   double *ximm_ccl_out2=malloc((nr-N1)*sizeof(double));
 
-  double z = j+0.;	
+  double z = j+0.;
   ccl_correlation_3dRsd_avgmu(cosmo,1.0,N1,r_arr1,beta[i_model-1],ximm_ccl_out1,&status);
   ccl_correlation_3dRsd_avgmu(cosmo,1.0,nr-N1,r_arr2,beta[i_model-1],ximm_ccl_out2,&status);
 
   if (status) printf("%s\n",cosmo->status_message);
-  for(i=0;i<nr;i++){     
+  for(i=0;i<nr;i++){
   double err;
 
   if(i<N1){
-  err=fabs(r_arr1[i]*r_arr1[i]*(ximm_ccl_out1[i]-ximm_bench_arr[i])); 
+  err=fabs(r_arr1[i]*r_arr1[i]*(ximm_ccl_out1[i]-ximm_bench_arr[i]));
   ASSERT_DBL_NEAR_TOL(0.,err,CORR_TOLERANCE1);
   }
   else{
-  err=fabs(r_arr2[i-N1]*r_arr2[i-N1]*(ximm_ccl_out2[i-N1]-ximm_bench_arr[i])); 
+  err=fabs(r_arr2[i-N1]*r_arr2[i-N1]*(ximm_ccl_out2[i-N1]-ximm_bench_arr[i]));
   ASSERT_DBL_NEAR_TOL(0.,err,CORR_TOLERANCE2);
   }
-  }   
+  }
   fclose(f);
-  
+
   free(r_arr1);
   free(r_arr2);
   free(ximm_bench_arr);
   free(ximm_ccl_out1);
-  free(ximm_ccl_out2); 
+  free(ximm_ccl_out2);
 
   ccl_cosmology_free(cosmo);
 }
