+++ tests/ccl_test_pyccl_interface.py
@@ -36,7 +36,7 @@ def reference_models():
     cosmo6 = ccl.Cosmology(
         Omega_c=0.27, Omega_b=0.022/0.67**2, h=0.67, sigma8=0.8,
         n_s=0.96, Neff=3.04, m_nu=0.,
-        transfer_function='emulator',
+        transfer_function='boltzmann_class',
         matter_power_spectrum='emu')
 
     # Baryons Pk
@@ -70,7 +70,7 @@ def reference_models_nu():
     cosmo1 = ccl.Cosmology(
         Omega_c=0.27, Omega_b=0.022/0.67**2, h=0.67, sigma8=0.8,
         n_s=0.96, Neff=3.04, m_nu=[0.02, 0.02, 0.02],
-        transfer_function='emulator',
+        transfer_function='boltzmann_class',
         matter_power_spectrum='emu')
 
     # Emulator Pk with neutrinos, force equalize
@@ -687,21 +687,6 @@ def check_corr_3dRSD(cosmo):
     assert_( all_finite(corr10))
     assert_( all_finite(corr11))
     assert_( all_finite(corr12))
-
-def test_valid_transfer_combos():
-    """
-    Check that invalid transfer_function <-> matter_power_spectrum pairs raise
-    an error.
-    """
-    params = { 'Omega_c': 0.27, 'Omega_b': 0.045, 'h': 0.67,
-               'A_s': 1e-10, 'n_s': 0.96, 'w0': -1., 'wa': 0. }
-
-    assert_raises(ValueError, ccl.Cosmology, transfer_function='emulator',
-                              matter_power_spectrum='linear', **params)
-    #assert_raises(ValueError, ccl.Cosmology, transfer_function='boltzmann',
-    #                          matter_power_spectrum='halomodel', **params)
-    assert_raises(ValueError, ccl.Cosmology, transfer_function='bbks',
-                              matter_power_spectrum='emu', **params)
 
 def test_background():
     """
